pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "This is a project package that contains an utility that determines the difference between two data structures"
authors = ["alllexxx1 <alekkorpachyov@gmail.com>"]
readme = "README.md"
repository = "https://github.com/alllexxx1/python-project-50.git"
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
	{ include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"